fixes for review

+++ uninstall.php
@@ -1,3 +1,5 @@
 <?php
+if ( ! defined( 'ABSPATH' ) ) exit; // Exit if accessed directly
+
 delete_option( 'dubco_selected_post_types' );
 delete_option( 'dubco_api_key' );
